+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - March 2023
+
+### Added
+
+* Integrated CDK into the solution to build CloudFormation templates
+
+### Changed
+
+* Updated Github Workflows
+
+### Removed
+
+* Removed existing CloudFormation templates from source/msam/build, source/events, and source/web-cloudformation directories
+
+### Contributors
+@dch90
+
 ## [1.11.0] - August 2022
 
 ### New